did-method-prism: rename $ cleanup
